Consistent error handling when streaming
The lhttpc_client process, spawned by the caller and linked to
it, used erlang:error/1 when an error was received from the socket.
This is not friendly for callers to deal with errors, as they
have to trap exits or monitor the lhttc_client process. It also
didn't respect the spec/documentation for lhttpc:get_body_part/1,2
because {error, Reason} could never be returned.